audit_events: executed at last, and the cost guard was right
The migration has carried "AUTHORED, NOT EXECUTED — every statement below is
unverified against a running DB" since 2026-08-07, which is why I have declined
to apply it all session. It has now been run, on a throwaway database on the
local Postgres 18 at :5433. Production was not touched.

IT APPLIES CLEANLY, and the contract holds:
  1 completeness   PASS — a write with no intent is still recorded, flagged
  2 rich event     PASS — intent honoured, actor carried
  3 changed_keys   PASS — the floor, no deep diff
  5 enforce raises PASS — service-role + no intent throws, in enforce mode
  7 retirement     PASS — the dormant table is renamed, not dropped
  C4a/b/c         PASS — direct INSERT/UPDATE/DELETE denied even to service_role
  R1–R5           PASS — the rollback fully reverses, including the rename

THE COST GUARD WAS RIGHT, and this is the finding. Measured per write:

  blob     baseline   with md5   without md5   md5 alone
  71 kB      21 ms     104 ms       37 ms        68 ms
  357 kB     99 ms     554 ms      209 ms       345 ms
  933 kB    433 ms    1852 ms      570 ms      1282 ms

md5(doc::text) serialises the whole JSONB before hashing, so it grows with the
blob and dominates everything else. The live store holds Laila CRM at 1.58 MB
and a 5.3 MB row — both PAST the worst case measured, which puts multiple
seconds on every save. So the migration now takes its own guard's advice:
fingerprints off, changed_keys kept. That drops the trigger's overhead at ~1 MB
from ~1419 ms to ~137 ms. A fingerprint only says THAT something moved, which
the event's existence already says; changed_keys says WHAT moved.

Two defects in the verification kit itself, both fixed:
· the bare-Postgres auth.uid() shim chokes on the empty claims string the
  scripts use to simulate service context — and check 5 PASSED on that error,
  catching my shim's exception instead of the raise it was testing for. Another
  green over the wrong thing. Now it passes on the real message.
· scratch_03 called pg_size_pretty on an integer, ambiguous on PG18. Cast.

Check 4 no longer asserts "before <> after" — it asserts the DECISION, and it
is two-sided: it fails if fingerprints come back, because that line decides
whether saving a programme is instant or takes three seconds, and re-enabling
should cost someone a deliberate re-measure.

STILL NOT APPLIED TO PRODUCTION. What was authorised was the migration as
committed; what is now verified is a modified one. That is the user's call, and
it is one command with the rollback block in hand.

Honest limit: C6 (owner-only RLS read) reports PASS but I connected as
superuser, which bypasses RLS — the script says so itself. RLS isolation is
NOT proven by this run.

+++ supabase/migrations/_verify/scratch_03_cost_measure.sql
@@ -89,7 +89,7 @@ do $$
 declare sz text; full_ms numeric; base_ms numeric; nofp_ms numeric;
 begin
   update public.adam_programs set data = public._gen_blob(200, 20, 200) where id='cost-1';
-  select pg_size_pretty(pg_column_size(data)) into sz from public.adam_programs where id='cost-1';
+  select pg_size_pretty(pg_column_size(data)::bigint) into sz from public.adam_programs where id='cost-1';
   full_ms := public._measure('cost-1', 20);
   execute 'alter table public.adam_programs disable trigger aura_audit_programs';
   base_ms := public._measure('cost-1', 20);
@@ -111,7 +111,7 @@ do $$
 declare sz text; full_ms numeric; base_ms numeric; nofp_ms numeric;
 begin
   update public.adam_programs set data = public._gen_blob(500, 40, 250) where id='cost-1';
-  select pg_size_pretty(pg_column_size(data)) into sz from public.adam_programs where id='cost-1';
+  select pg_size_pretty(pg_column_size(data)::bigint) into sz from public.adam_programs where id='cost-1';
   full_ms := public._measure('cost-1', 10);
   execute 'alter table public.adam_programs disable trigger aura_audit_programs';
   base_ms := public._measure('cost-1', 10);
@@ -133,7 +133,7 @@ do $$
 declare sz text; full_ms numeric; base_ms numeric; nofp_ms numeric;
 begin
   update public.adam_programs set data = public._gen_blob(800, 60, 300) where id='cost-1';
-  select pg_size_pretty(pg_column_size(data)) into sz from public.adam_programs where id='cost-1';
+  select pg_size_pretty(pg_column_size(data)::bigint) into sz from public.adam_programs where id='cost-1';
   full_ms := public._measure('cost-1', 6);
   execute 'alter table public.adam_programs disable trigger aura_audit_programs';
   base_ms := public._measure('cost-1', 6);
